Split out the different checkers into their own files.
